knockAI/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
openai==1.54.3
python-dotenv==1.0.1
apscheduler==3.10.4
pandas==2.2.3
numpy>=2.0.0,<2.1.0
yfinance==0.2.47
python-multipart==0.0.12



